% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRawDataCommands.r
\name{convertRawDataCommands}
\alias{convertRawDataCommands}
\title{convertRawDataCommands}
\usage{
convertRawDataCommands(pathToRawData, whereFilesCreated,
  msConvertFolderLocation)
}
\arguments{
\item{pathToRawData}{directory that contains raw LC-MS/MS data}

\item{whereFilesCreated}{directory that will contain the new mzXML files}

\item{msConvertFolderLocation}{directory that contains msconvert.exe}
}
\value{
List of cmd.exe calls, each element is a single LC-MS/MS file to be converted
}
\description{
This function returns a list of cmd.exe calls to msconvert.exe. Note: doesn't make the calls, just the text to do so.
}
