% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMSConvert.r
\name{findpwiz}
\alias{findpwiz}
\title{findpwiz}
\usage{
findpwiz(userGivenPath, proteoWizardLocation)
}
\arguments{
\item{userGivenPath}{directory that contains "Specify_ProteoWizard_Location.txt"}

\item{proteoWizardLocation}{directory that contains msconvert.exe}
}
\value{
filepath of the folder containing msconvert.exe
}
